requirements.txt
# === FILE: requirements.txt ===
streamlit
simple-salesforce
python-arango
pandas